split AddUser panel out into separate component
